+++ base/src/main/java/com/avogine/game/scene/Projection.java
@@ -3,20 +3,22 @@ package com.avogine.game.scene;
 import org.joml.Matrix4f;
 
 /**
- *
+ * TODO Add methods to update just the FOV/NEAR/FAR
  */
 public class Projection {
 	
 	private static final float DEFAULT_FOV = (float) Math.toRadians(60f);
 	private static final float DEFAULT_Z_NEAR = 0.01f;
 	private static final float DEFAULT_Z_FAR = 1000.0f;
-	
-	private final float fov;
-	private final float zNear;
-	private final float zFar;
-	
+
 	private final Matrix4f projectionMatrix;
 	private final Matrix4f invertedProjectionMatrix;
+	
+	private float fov;
+	private int width;
+	private int height;
+	private float zNear;
+	private float zFar;
 	
 	/**
 	 * @param fov 
@@ -26,12 +28,14 @@ public class Projection {
 	 * @param zFar 
 	 */
 	public Projection(float fov, int width, int height, float zNear, float zFar) {
-		this.fov = fov;
 		projectionMatrix = new Matrix4f();
 		invertedProjectionMatrix = new Matrix4f();
+		this.fov = fov;
+		this.width = width;
+		this.height = height;
 		this.zNear = zNear;
 		this.zFar = zFar;
-		updateProjectionMatrix(width, height);
+		setAspectRatio(width, height);
 	}
 	
 	/**
@@ -42,6 +46,22 @@ public class Projection {
 		this(DEFAULT_FOV, width, height, DEFAULT_Z_NEAR, DEFAULT_Z_FAR);
 	}
 	
+	/**
+	 * This will not modify the projection matrix, so should be the preferred way to retrieve an inverted result of this projection matrix.
+	 * @return the result of calling {@link Matrix4f#invert()} on {@link #getProjectionMatrix()}.
+	 */
+	public Matrix4f invert() {
+		return projectionMatrix.invert(invertedProjectionMatrix);
+	}
+	
+	/**
+	 * @param width
+	 * @param height
+	 */
+	public void setAspectRatio(int width, int height) {
+		projectionMatrix.setPerspective(fov, (float) width / height, zNear, zFar);
+	}
+
 	/**
 	 * @return the projectionMatrix
 	 */
@@ -57,19 +77,17 @@ public class Projection {
 	}
 	
 	/**
-	 * This will not modify the projection matrix, so should be the preferred way to retrieve an inverted result of this projection matrix.
-	 * @return the result of calling {@link Matrix4f#invert()} on {@link #getProjectionMatrix()}.
+	 * @return the aspect
 	 */
-	public Matrix4f invert() {
-		return projectionMatrix.invert(invertedProjectionMatrix);
+	public int getWidth() {
+		return width;
 	}
 	
 	/**
-	 * @param width
-	 * @param height
+	 * @return the height
 	 */
-	public void updateProjectionMatrix(int width, int height) {
-		projectionMatrix.setPerspective(fov, (float) width / height, zNear, zFar);
+	public int getHeight() {
+		return height;
 	}
 	
 }
