+++ base/src/main/java/com/avogine/game/ui/GUI.java
@@ -1,0 +1,8 @@
+package com.avogine.game.ui;
+
+/**
+ *
+ */
+public interface GUI {
+
+}
