+++ base/src/main/java/com/avogine/game/ui/UIElement.java
@@ -1,0 +1,16 @@
+package com.avogine.game.ui;
+
+import com.avogine.io.Window;
+
+/**
+ * @param <T> 
+ */
+public interface UIElement<T> {
+
+	/**
+	 * @param context
+	 * @param window 
+	 */
+	public void layout(T context, Window window);
+	
+}
