--- base/src/main/java/com/avogine/game/ui/nuklear/UIElement.java
@@ -1,27 +1,0 @@
-package com.avogine.game.ui.nuklear;
-
-import org.lwjgl.nuklear.NkContext;
-
-import com.avogine.io.Window;
-
-/**
- * TODO Further subclass this into a NuklearUIElement for prepare to take in specifically a NkContext
- */
-public interface UIElement {
-
-	/**
-	 * TODO Maybe pass in the window so elements can dynamically size/translate themselves
-	 * Might also be worth making this private and forcing implementation of Renderable
-	 * @param context
-	 * @param windowId 
-	 */
-	public void prepare(NkContext context, Window window);
-	
-	/**
-	 * @return whether this UI should immediately be rendered once initialized, otherwise it will start hidden.
-	 */
-	public default boolean showOnInit() {
-		return true;
-	}
-	
-}
