+++ base/src/main/java/com/avogine/game/util/Registerable.java
@@ -1,0 +1,15 @@
+package com.avogine.game.util;
+
+import com.avogine.io.Window;
+
+/**
+ * Interface for game components that need additional initialization that must occur on the main thread.
+ */
+public interface Registerable {
+
+	/**
+	 * @param window 
+	 */
+	public void init(Window window);
+	
+}
