+++ base/src/main/java/com/avogine/io/event/KeyEvent.java
@@ -5,36 +5,43 @@ import java.util.concurrent.atomic.AtomicBoolean;
 import com.avogine.io.listener.KeyListener;
 
 /**
- * Base interface for {@link InputEvent}s fired by key input.
+ * Base interface for {@link InputEvent}s fired by keyboard input.
  */
 public sealed interface KeyEvent extends InputEvent, ConsumableEvent {
 	/**
-	 * @return the key
+	 * @return the keyboard key that was pressed or released.
 	 */
 	public int key();
 	
 	/**
-	 * @return the unicode codepoint
+	 * @return the platform-specific scancode of the key.
 	 */
-	public int codepoint();
+	public int scancode();
+	
+	/**
+	 * @return bitfield describing which modifiers keys were held down.
+	 */
+	public int mods();
 
 	/**
 	 * KeyEvent where a key is pressed.
 	 * <p>
 	 * This will continue to fire new events while the key is pressed.
 	 * @param window The window ID the key was triggered from.
-	 * @param key The key that triggered the event
-	 * @param codepoint The character produced by this event.
+	 * @param key the keyboard key that was pressed.
+	 * @param scancode the platform-specific scancode of the key.
+	 * @param mods Bitfield describing which modifiers keys were held down.
 	 * @param consumed true only if this event has been handled and consumed by a {@link KeyListener}.
 	 */
-	public record KeyPressedEvent(long window, int key, int codepoint, AtomicBoolean consumed) implements KeyEvent {
+	public record KeyPressedEvent(long window, int key, int scancode, int mods, AtomicBoolean consumed) implements KeyEvent {
 		/**
 		 * @param window The window ID the key was triggered from.
 		 * @param key The key that triggered the event
-		 * @param codepoint The character produced by this event.
+		 * @param scancode The platform-specific scancode of the key.
+		 * @param mods Bitfield describing which modifiers keys were held down.
 		 */
-		public KeyPressedEvent(long window, int key, int codepoint) {
-			this(window, key, codepoint, new AtomicBoolean());
+		public KeyPressedEvent(long window, int key, int scancode, int mods) {
+			this(window, key, scancode, mods, new AtomicBoolean());
 		}
 		
 		@Override
@@ -46,18 +53,20 @@ public sealed interface KeyEvent extends InputEvent, ConsumableEvent {
 	/**
 	 * KeyEvent where a key was released.
 	 * @param window The window ID the key was triggered from.
-	 * @param key The key that triggered the event
-	 * @param codepoint The character produced by this event.
+	 * @param key The keyboard key that was released.
+	 * @param scancode The platform-specific scancode of the key.
+	 * @param mods Bitfield describing which modifiers keys were held down.
 	 * @param consumed true only if this event has been handled and consumed by a {@link KeyListener}.
 	 */
-	public record KeyReleasedEvent(long window, int key, int codepoint, AtomicBoolean consumed) implements KeyEvent {
+	public record KeyReleasedEvent(long window, int key, int scancode, int mods, AtomicBoolean consumed) implements KeyEvent {
 		/**
 		 * @param window The window ID the key was triggered from.
 		 * @param key The key that triggered the event
-		 * @param codepoint The character produced by this event.
+		 * @param scancode The character produced by this event.
+		 * @param mods Bitfield describing which modifiers keys were held down.
 		 */
-		public KeyReleasedEvent(long window, int key, int codepoint) {
-			this(window, key, codepoint, new AtomicBoolean());
+		public KeyReleasedEvent(long window, int key, int scancode, int mods) {
+			this(window, key, scancode, mods, new AtomicBoolean());
 		}
 		
 		@Override
@@ -66,16 +75,4 @@ public sealed interface KeyEvent extends InputEvent, ConsumableEvent {
 		}
 	}
 	
-	/**
-	 * KeyEvent where a key was just pressed.
-	 * <p>
-	 * This differs from {@link KeyPressedEvent} in that it will only be fired once per key press.
-	 * @param window The window ID the key was triggered from.
-	 * @param key The key that triggered the event
-	 * @param codepoint The character produced by this event.
-	 */
-	public record KeyTypedEvent(long window, int key, int codepoint) implements KeyEvent {
-		
-	}
-
 }
