+++ base/src/main/java/com/avogine/io/listener/CharListener.java
@@ -1,0 +1,15 @@
+package com.avogine.io.listener;
+
+import com.avogine.io.event.CharEvent;
+
+/**
+ *
+ */
+public interface CharListener extends InputListener {
+
+	/**
+	 * @param event
+	 */
+	public void charTyped(CharEvent event);
+	
+}
